Fixed the way Signal inputs and outputs are handled when instrumenting files in tests.
Previously, there was an issue due to the fact Angular's `input()` and `output()` functions cannot be used outside the initialiser of a class, but in order to add coverage to class attributes, Istanbul's coverage function call must be called before the value assigned to the attribute. So while input and output signals can only be used as `property = input()`, Istanbul's instrumentation converted the code to `property = (cov_func, input()). This broke inputs and outputs as Angular's compiler no longer recognised them as component inputs and outputs, but rather as regular properties. The instrumentation is now handled slightly differently for inputs and outputs - rather than being instrumented before the Angular Compiler compiles the code, their instrumentation is handled after the code has been compiled, which ensures both Angular's and Istanbul's functionality operates correctly.